# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Baqaya API",
        "description": "Production API for Baqaya — offline-first ledger, customers, payments, sync, statements, and reminders.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Baqaya API Server (same host as this documentation)"
        }
    ],
    "paths": {
...
        "/api/v1/shop": {
            "get": {
                "tags": [
                    "Shop"
                ],
                "summary": "Get authenticated user's shop",
                "description": "Returns the shop owned by the authenticated user. No shop UUID header is required.",
                "operationId": "shopShow",
                "responses": {
                    "200": {
                        "description": "The user's shop",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Shop"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Shop not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Shop"
                ],
                "summary": "Create shop",
                "description": "Creates the authenticated user's shop. Each user may own exactly one shop.",
                "operationId": "shopStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Arslan Kiryana"
                                    },
                                    "business_type": {
                                        "type": "string",
                                        "example": "grocery"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+923001111111"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "PKR"
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Asia/Karachi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shop created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Shop"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or shop already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Shop"
                ],
                "summary": "Soft-delete authenticated user's shop",
                "operationId": "shopDestroy",
                "responses": {
                    "200": {
                        "description": "Shop deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Shop"
                ],
                "summary": "Update authenticated user's shop",
                "description": "Updates the shop owned by the authenticated user.",
                "operationId": "shopUpdate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Shop updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/ApiSuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/Shop"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Shop": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "01JXYZABCDEFGHJKMNPQRSTVW"
                    },
                    "name": {
                        "type": "string",
                        "example": "Arslan Kiryana"
                    },
                    "business_type": {
                        "type": "string",
                        "example": "grocery",
                        "nullable": true
                    },
                    "phone": {
                        "type": "string",
                        "example": "+923001111111",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "currency": {
                        "type": "string",
                        "example": "PKR"
                    },
                    "timezone": {
                        "type": "string",
                        "example": "Asia/Karachi"
                    },
                    "server_version": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "ApiSuccessResponse": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "description": "Response payload (shape varies by endpoint)"
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "last_page": {
                                "type": "integer",
                                "example": 3
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 20
                            },
                            "total": {
                                "type": "integer",
                                "example": 42
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiErrorResponse": {
                "required": [
                    "success",
                    "message",
                    "code"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    },
                    "code": {
                        "type": "string",
                        "example": "NOT_FOUND"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true
                    },
                    "meta": {
                        "description": "Extra context (e.g. sync conflict)",
                        "properties": {
                            "latest_server_version": {
                                "type": "integer",
                                "example": 123
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    },
                    {
                        "properties": {
                            "code": {
                                "type": "string",
                                "example": "VALIDATION_ERROR"
                            },
                            "errors": {
                                "type": "object",
                                "example": {
                                    "phone": [
                                        "The phone field is required."
                                    ]
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UnauthorizedResponse": {
                "example": {
                    "success": false,
                    "message": "Unauthenticated.",
                    "code": "UNAUTHENTICATED"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                ]
            },
            "NotFoundResponse": {
                "example": {
                    "success": false,
                    "message": "Resource not found.",
                    "code": "NOT_FOUND"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ApiErrorResponse"
                    }
                ]
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste the access_token from OTP verify (Bearer prefix optional)",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}